EditLabForm -> LabForm, добавление проверки 400, разделение модальных окон
