--- src/styles/components/EditLabForm.module.scss
@@ -1,21 +1,0 @@
-.container {
-    display: flex;
-    flex-direction: column;
-    gap: 16px;
-    padding: 21px;
-
-    & .submitButton {
-        border-radius: 6px;
-        border: 1px;
-        width: 140px;
-        height: 45px;
-
-        align-self: center;
-        margin-left: auto;
-        width: max-content;
-
-        font-size: 20px;
-        font-weight: 500;
-        line-height: 24.2px;
-    }
-}
