+++ src/styles/components/LabForm.module.scss
@@ -2,29 +2,20 @@
     display: flex;
     flex-direction: column;
     gap: 16px;
-  
+    padding: 21px;
+
     & .submitButton {
+        border-radius: 6px;
+        border: 1px;
+        width: 140px;
+        height: 45px;
+
+        align-self: center;
+        margin-left: auto;
         width: max-content;
-    }
 
-    & .calendar {
-        display: flex;
-        flex-direction: column;
-        gap: 8px;
-    }
-
-    & .buttonContainer {
-        display: flex;
-        justify-content: flex-end;
-    }
-
-    & .submitButton {
-        padding: 10.5px 17.5px;
-        margin-block-start: 26px;
-        font-family: Inter, sans-serif;
         font-size: 20px;
         font-weight: 500;
         line-height: 24.2px;
-        text-align: left;
-       }
-  }
+    }
+}
